Keep DeletedFilter when clearing list of modifiers
The DeletedFilter is needed to filter out deleted annotations. This
has to do with the way elasticsearch deletes documents in it's index.
Rather than removing the deleted document from the disk, elasticsearch
marks documents as deleted which means unless they are filtered out,
they will still be returned in a search until a flush happens and the
documents marked as deleted are removed from disk.