When deleting annotations refresh all shards
Elasticsearch by default has 5 shards that it searches from. When
a delete is issued from delete_by_query it only marks the annotation
as deleted in the current shard. In order to make sure that change
propogates to all the shards refresh must be set to True.

A version conflict can occur when a document has been updated but
the update has not been propogated to all shards; meaning the current
shard has a lower version number than a different shard that updated
the document. The comment in the code was updated to explain the
reason why this error was happening.